Removed the old collision callback system and replaced it with a much simpler and more efficient iterate-over-PersistentManifolds Updated PInvoke with latest changes from Upstream repository Updated iOS library

+++ UnityProject/Assets/Plugins/BulletSharp/Extras/Chunk.cs
@@ -64,7 +64,7 @@ namespace BulletSharp
         {
             Code = (DnaID)c.Code;
             Length = c.Length;
-            OldPtr = c.UniqueInt1 + c.UniqueInt2 << 32;
+            OldPtr = c.UniqueInt1 + ((long)c.UniqueInt2 << 32);
             DnaNR = c.DnaNR;
             NR = c.NR;
         }
